backend/requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
google-generativeai>=0.3.0
python-dotenv>=1.0.0
